// src/app/interfaces/player.ts
export interface Player {
  id?: string;
  name: string;
  battingAvg: number;
  gamesPlayed: number;
  plateAppearances: number;
  atBats: number;
  runs: number;
  hits: number;
  doubles: number;
  triples: number;
  homeRuns: number;
  rbis: number;
  walks: number;
  strikeouts: number;
  sacrificeFlies: number;
  totalBases: number;
  obp: number;
  slg: number;
  ops: number;
}